requirements.txt
PyMuPDF
unidecode
langdetect
polyglot
pdfminer.six
langcodes
pytesseract
pillow
